moved tests, removed conn stubs
